api +cronjob

+++ src/main/java/com/example/BookExchange/dto/BooksDto.java
@@ -8,9 +8,9 @@ public class BooksDto {
     private String name;
     private String categorie;
     private String image;
-    private Long prix;
+    private Double prix;
 
-    public BooksDto(String name, String categorie, String image, Long prix) {
+    public BooksDto(String name, String categorie, String image, Double prix) {
         this.name = name;
         this.categorie = categorie;
         this.image = image;
